Move reader into src and delete old location
